requirements.txt
fastapi==0.110.2
uvicorn==0.29.0
websockets==11.0.3
redis==4.6.0
httpx==0.27.0
pydantic==2.7.1
python-dotenv==1.0.1
python-logging-loki==0.3.1